wip/gcp/requirements.txt
google-cloud-bigquery [bqstorage,pandas]
pandas-gbq
apache-beam
